Sorta working with customisable ui terriajs.
