--- TerriaMap.scss
@@ -1,12 +1,0 @@
-@import "~terriajs/lib/Sass/StandardUserInterface";
-
-.panel--getting-started .panel-content{
-  background-image: url("~terriajs/wwwroot/images/welcome-bg.jpg");
-}
-
-@media print {
-    .nationalmap-print-disclaimer {
-        color: #000;
-        display: block;
-    }
-}
