--- lib/Views/RelatedMaps.html
@@ -1,50 +1,0 @@
-<div class="nationalmap-related-maps-intro">
-    Clicking on a map below will open it in a separate window or tab.
-</div>
-<table class="nationalmap-related-maps-table">
-    <tr>
-        <td>
-            <a href="http://nationalmap.gov.au/renewables/">
-                <img src="images/aremi.jpg" width="200" height="150" alt="AREMI" />
-            </a>
-        </td>
-        <td>
-            <h2><a href="http://nationalmap.gov.au/renewables/">AREMI</a></h2>
-            AREMI provides access to Australian spatial data relevant to the Renewable Energy industry, sourced from Government, Industry and Research.
-        </td>
-    </tr>
-    <tr>
-        <td>
-            <a href="http://nationalmap.gov.au/northernaustralia/">
-                <img src="images/northernaustralia.jpg" width="200" height="150" alt="Northern Australia" />
-            </a>
-        </td>
-        <td>
-            <h2><a href="http://nationalmap.gov.au/northernaustralia/">Northern Australia</a></h2>
-            The Northern Australia NationalMap forms part of the Government's commitment to developing northern Australia by providing easy access to authoritative and other spatial data on northern Australia to governments, business and the public.
-
-More information on the Government's White Paper on Developing Northern Australia is available here:  https://northernaustralia.dpmc.gov.au
-        </td>
-    </tr>
-    <tr>
-        <td>
-            <a href="http://neiiviewer.nicta.com.au">
-                <img src="images/neii.jpg" width="200" height="150" alt="NEII Viewer" />
-            </a>
-        </td>
-        <td>
-## [NEII Viewer](http://neiiviewer.nicta.com.au)
-The National Environmental Information Infrastructure (NEII) is an information platform designed to improve discovery, access and re-use of nationally significant environmental data. More information on the NEII is available here: [www.neii.gov.au](http://www.neii.gov.au)</td>
-    </tr>
-    <tr>
-        <td>
-            <a href="http://map.aurin.org.au">
-                <img src="images/aurin-map.jpg" width="200" height="150" alt="AURIN Map" />
-            </a>
-        </td>
-        <td>
-## [AURIN Map](http://map.aurin.org.au)
-AURIN Map provides access to datasets on urban infrastructure for urban researchers, policy and decision makers.
-    </td>
-    </tr>
-</table>
