+++ lib/Views/global.scss
@@ -1,0 +1,12 @@
+@import "~terriajs/lib/Sass/StandardUserInterface";
+
+.panel--getting-started .panel-content {
+	background-image: url("~terriajs/wwwroot/images/welcome-bg.jpg");
+}
+
+@media print {
+	.nationalmap-print-disclaimer {
+		color: #000;
+		display: block;
+	}
+}
