+++ wwwroot/init/test.json
@@ -444,6 +444,44 @@
       ]
     },
     {
+      "name": "Terria Analytics",
+      "type": "group",
+      "preserveOrder": true,
+      "items": [
+        {
+          "name": "Regions like this",
+          "type": "places-like-me-function",
+          "url": "http://parlmap.terria.io:31113/jobs"
+        },
+        {
+          "name": "What makes this region unique or special?",
+          "type": "why-am-i-special-function",
+          "url": "http://parlmap.terria.io:31113/jobs"
+        },
+        {
+          "name": "Spatial Detailing",
+          "type": "spatial-detailing-function",
+          "url": "http://parlmap.terria.io:31113/jobs"
+        },
+        {
+          "name": "Sample Data",
+          "type": "group",
+          "items": [
+            {
+              "name": "NSW NEXIS 2012 (LGA)",
+              "type": "csv",
+              "url": "test/NSW_LGA_NEXIS_201212.csv"
+            },
+            {
+              "name": "Victoria Crime (LGA)",
+              "type": "csv",
+              "url": "test/vic_assault.csv"
+            }
+          ]
+        }
+      ]
+    },
+    {
       "name": "Empty Group",
       "type": "group"
     }
